+++ test/OpenTelemetry.OpAmp.Client.Tests/FrameProcessorTests.cs
@@ -13,12 +13,12 @@ public class FrameProcessorTests
     [Fact]
     public void FrameProcessor_Subscribe()
     {
+        using var listener = new MockListener();
         var processor = new FrameProcessor();
-        var listener = new MockListener();
-        var mockFrame = FrameGenerator.GenerateMockFrame();
+        var mockFrame = FrameGenerator.GenerateMockServerFrame();
 
         processor.Subscribe(listener);
-        processor.OnServerFrame(mockFrame.Frame);
+        processor.OnServerFrame(mockFrame.Frame.ToSequence());
 
         var message = Assert.Single(listener.Messages);
         Assert.Equal(mockFrame.ExptectedContent, message.CustomMessage.Data.ToStringUtf8());
@@ -27,26 +27,26 @@ public class FrameProcessorTests
     [Fact]
     public void FrameProcessor_Unsubscribe()
     {
+        using var listener = new MockListener();
         var processor = new FrameProcessor();
-        var listener = new MockListener();
-        var mockFrame = FrameGenerator.GenerateMockFrame();
+        var mockFrame = FrameGenerator.GenerateMockServerFrame();
 
         processor.Subscribe(listener);
-        processor.OnServerFrame(mockFrame.Frame);
+        processor.OnServerFrame(mockFrame.Frame.ToSequence());
 
         Assert.Single(listener.Messages);
 
         processor.Unsubscribe(listener);
-        processor.OnServerFrame(mockFrame.Frame);
+        processor.OnServerFrame(mockFrame.Frame.ToSequence());
         Assert.Single(listener.Messages);
     }
 
     [Fact]
     public async Task FrameProcessor_ThreadSafety()
     {
+        using var listener = new MockListener();
         var processor = new FrameProcessor();
-        var listener = new MockListener();
-        var mockFrame = FrameGenerator.GenerateMockFrame();
+        var mockFrame = FrameGenerator.GenerateMockServerFrame();
         int iterations = 1000;
         var tasks = new List<Task>
         {
@@ -55,7 +55,7 @@ public class FrameProcessorTests
             {
                 Parallel.For(0, iterations, i =>
                 {
-                    processor.OnServerFrame(mockFrame.Frame);
+                    processor.OnServerFrame(mockFrame.Frame.ToSequence());
                 });
             }),