+++ src/OpenTelemetry.OpAmp.Client/CHANGELOG.md
@@ -6,3 +6,5 @@
   ([#2917](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2917))
 * Added support for OpAMP Plain HTTP transport.
   ([#2926](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2926))
+* Added support for OpAMP WebSocket transport.
+  ([#2926](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3064))